# TINYGL: Resample texture during rendering