~~Note: based on #1339.~~

Before this change, TINYGL resizes textures when they are uploaded to GL (`tglTexImage2D`) to the init-configured internal square texture size, and applies bilinear filtering at that time.

This has 2 effects:
- memory: most textures are smaller than (currently) requested internal texture size, so they use more memory once sent to TinyGL
- rendering: interpolation happens browsing original pixel grid by walking the resized pixel grid with borders adjusted to line up with the original image. Resulting texture will show re-sampling artifacts when rendered if it is not aligned to screen (...which is mostly visible for high-resolution original images)

Moving bilinear filtering to render time has potentially a large cost: it needs to read 3 texels for each rendered pixel, and reading texture already requires expensive fetches from RAM.

But this cost can be significantly reduced by relying on CPU data cache: each fetch from main memory actually loads several adjacent words in CPU cache, so accessing nearby bytes is cheap.

This is achieved by repacking the original texture, bringing the 4 possible involved texels (of which 3 will be used, choosing which depending on sampling offset) close to each other.

This again has an obvious cost: each texture takes 4 times its original size. But this cost is offset by being working on texture in its original size, saving lots of ram on smaller textures and should allow increasing internal texture resolution with less negative effects than before (but I did not try this much).

# Functional effect

This change allows implementing simply:

- `TGL_NEAREST` texture filtering (although texture sampler is still chosen at texture upload time, not at render time - I did not check what OpenGL is doing, but at least as of current MYST3 and GRIM engines this assumption is correct)
- `TGL_TEXTURE_WRAP_`{`S`,`T`}

# Visual effect
- ship textures in EMI (very visible in first puzzle, more subtle in shipyard)
- myst becomes a lot sharper, and I feel runs even much smoother. As it is uploading textures very often, old TinyGL code was resampling large (640x640) textures to 512x512 a lot, and probably often in vain as their polygon could be fully off-screen.

# Memory usage effect

[EDIT] Measures updated after fixing a per-`tglTexImage2D`-call memory leak (mostly affecting MYST3)

Game | Set | Before | After, `TGL_LINEAR` | After, `TGL_NEAREST`
---- | --- | ------ | ----------------- | ------------------
GRIM | cn | 78 MB | 67 MB | 58 MB
EMI | first puzzle | 91 MB | 91 MB | 81 MB
MYST3 | irrelevant ? | 82 MB | 103 MB | 71 MB

GRIM is an easy win, many textures are smaller than 256x256, so they were made needlessly large before.
EMI first puzzle is a tough one, as the moving ship in the background is a large texture.
Lastly, MYST is full of high-rez textures, larger than its requested 512x512 internal texture size, so it is a loss. But I believe it is a reasonable one compared to the visual and performance improvement.

# Performance effect

Despite the texel repacking trick, this has a mild negative impact on FPS, especially when disabling dirtyrects.

Here are my measurments, with `--engine-speed=0 --no-dirtyrects`

Game | Set | Before | After
---- | --- | ------ | -----
GRIM | mo | 540 fps | 510 fps
GRIM | cn | 185 fps | 160 fps
GRIM | bi | 165 fps | 150 fps
EMI | first puzzle | 110 fps | 78 fps

I believe this performance decrease to be reasonable. But I'm biased :) .

[EDIT] Not based on #1339 anymore