@vpelletier 2017-05-12
MYST opening scene

Before (ignore the black dot in the sky, it's the bird moving around):
![image](https://cloud.githubusercontent.com/assets/191204/26000438/7b256558-3718-11e7-8ae5-f0d655428273.png)

After:
![image](https://cloud.githubusercontent.com/assets/191204/26000449/89a86a76-3718-11e7-8c92-3cd4b8fdec74.png)