@vpelletier 2017-05-13
Changes since last push:
- rebase on latest 1339
- fix per-`tglTexImage2D`-call memory leak