@bgK 2017-05-13
The visual change is impressive :sunglasses: 

I did a quick performance test with Myst III (LEIS 3):
* master: 41 fps
* tinygl_texture_pixelformat (#1339): 40 fps
* tinygl_texture_sampler (#1343): 30 fps