@vpelletier 2017-05-13
Aha, so it's actually slower - and quite a bit.

How did you measure ? I did not see code for it in MYST3. I'ts very helpful (for me) in GRIM, so maybe it should be controlled by --fps in MYST3 too.