@bgK 2017-05-13
TinyGL gets patched through OpenGL by SDL2 on my system. I use MESA's features to measure the FPS:
VBLANK_MODE=0 LIBGL_SHOW_FPS=1 ./residualvm