@vpelletier 2017-05-13
> I use MESA's features to measure the FPS:

Nvidia blob here, so no luck.