@vpelletier 2017-05-13
This PR introduces a one-pixel offset (or scaling error ?) on EMI save-game screenshots.

These are a bit special: the texture is much larger than the displayed portion and is bordered by solid color (apparently white for the "broken save" and "empty slot" images, black for actual screenshots).