@vpelletier 2017-07-20
Changes since last push:
- made independent from #1339, which forces disabling endian-dependent pixel formats
- fixed a typo in `#if defined(TINYGL_LITLE_ENDIAN)` (was missing one `T`), although it is now disabled (see above)

I would be interested by:
- opinion whether this should go in, feature-wise, despite the performance impact (or maybe I should make TinyGL just not request linear filtering ?)
- opinion on the naming of the new file and classes in tinygl
- tests that OpenGL does not regress following this patch set (which should only be related to 	`GRAPHICS: Store pixels in native order in PixelBuffer`), especially on big-endian as I still fail to get reliable results even with linux on native hardware (I should set Xcode up and build residualvm there eventually). I'm not interested in stuff that was and is still broken, although stuff that would not be broken the same way could be interesting.