@bgK 2017-07-30
* Regarding `GRAPHICS: Store pixels in native order in PixelBuffer`, I very much agree PixelBuffer should not have to deal with endianness. I have no means to test the changes on big endian though.
* Regarding Myst III: The textures for the sides of the cube are 640x640. They are fully updated by the water effect. The other texture updates (mostly spot items) only update the affected portion of the cube sides.

I did not look at the code in too much detail since I only have a very shallow understanding of how TinyGL works, but the changes look to me good overall.