@vpelletier 2017-08-15
Thanks for the review.

For Myst III, indeed water effect costs a lot. I noticed this when facing the "rainbow" door (a handy color format test :) ) on Jnanin: when the door is on the left border of the screen, framerate goes up a lot as we are likely facing a no-sea face, while all 3 other faces have waves.

Also, I just realised wave effect can be disabled, so I think this should not be a big blocker.

Profiling some more, I find that, when using TGL_NEAREST, the hot-spot (30% of cpu time) is PixelBuffer's color format conversion: my framebuffer is (from largest LE shift to lowest, IOW as printed by `ColorFormat.asText`): `_RGB` (`_` because the format is 4 bytes per pixel, despite having no alpha channel), but MYST3 passes textures as `ABGR`, so byte-swapping has to happen.

Tweaking texture formats so no byte-swapping is needed, and making PixelBuffer ignore component difference when going from 1+ bits to 0 bits at same shift accelerates significantly.

But I believe this is a separate improvement, which may need graphics layer to expose its preferred pixel format to engines, and then engines to use it (...or not, if not hot-spot, like myst3 menu textures) when converting textures on load.

So I am very tempted to merge this PR.