@vpelletier 2017-08-15
Rebased on latest master, plus 2 commits:
- I used convoluted fixed-point code to try to avoid floating-point operations for coordinates conversion in texture sampler. But it is just slow, and trivial code using floating-point is faster (at least on my machine). I add it in a separate commit, but I am tempted to squash in the commit adding texture sampler.
- I also make the change to PixelBuffer format conversion I just talked about: it will try harder to use the fast memcpy path. No change to engines, so for example Myst3 will not take advantage of this on my machine.